Move `query-highlight-store` to `models` feature dir
