+++ web-local/src/lib/application-state-stores/initialize-node-store-contexts.ts
@@ -1,7 +1,7 @@
 /*global  RILL_RUNTIME_URL */
 import { browser } from "$app/environment";
+import { createQueryHighlightStore } from "@rilldata/web-common/features/models/query-highlight-store";
 import { createApplicationBuildMetadataStore } from "@rilldata/web-local/lib/application-state-stores/build-metadata";
-import { createQueryHighlightStore } from "@rilldata/web-local/lib/application-state-stores/query-highlight-store";
 import { setContext } from "svelte";
 
 /** determined by Vite's define option. */
